clip

+++ src/deepcsf/models/model_csf.py
@@ -80,9 +80,9 @@ class ContrastDiscrimination(CSFNetwork):
         x0 = self.check_img_type(x0)
         x1 = self.check_img_type(x1)
         x0 = self.features(x0)
-        x0 = x0.view(x0.size(0), -1)
+        x0 = x0.view(x0.size(0), -1).float()
         x1 = self.features(x1)
-        x1 = x1.view(x1.size(0), -1)
+        x1 = x1.view(x1.size(0), -1).float()
         x = torch.cat([x0, x1], dim=1)
         x = self.fc(x)
         return x
@@ -95,6 +95,6 @@ class GratingDetector(CSFNetwork):
     def forward(self, x):
         x = self.check_img_type(x)
         x = self.features(x)
-        x = x.view(x.size(0), -1)
+        x = x.view(x.size(0), -1).float()
         x = self.fc(x)
         return x
